// web3/blockchain/enum/query.names.enum.ts
export enum QueryNames {
  ALL_COLLECTIONS = "ALL_COLLECTIONS",
  HOT_NFTS = "HOT_NFTS",
  COLLECTIONS_BY_CATEGORIES = "COLLECTIONS_BY_CATEGORIES",
  ALL_NFTS = "ALL_NFTS",
  NFTS_BY_CATEGORY = "NFTS_BY_CATEGORY",
  NFT = "NFT",
  SALE_HISTORY = "SALE_HISTORY",
  COLLECTION = "COLLECTION",
  COLLECTION_NFTS = "COLLECTION_NFTS",
  NFT_BY_SALE_STATE = "NFT_BY_SALE_STATE",
  ACCOUNT_COLLECTION = "ACCOUNT_COLLECTION",
  LISTED_NFT_BY_ACCOUNT = "LISTED_NFT_BY_ACCOUNT",
  LISTED_USER_NFT_BY_ACCOUNT = "LISTED_USER_NFT_BY_ACCOUNT",
  ACCOUNT_CREATED_NFTS = "ACCOUNT_CREATED_NFTS",
  REGISTERED_COLLECTION = "REGISTERED_COLLECTION",
  COLLECTIONS_BY_ACCOUNT = "COLLECTIONS_BY_ACCOUNT",
  TOP_CREATOR_QUERY = "TOP_CREATOR_QUERY",
  GENEALOGY = "GENEALOGY",
  GENEALOGY_AT_LEVEL = "GENEALOGY_AT_LEVEL",
  REFERRAL_REWARD_IN_PRESALE = "REFERRAL_REWARD_IN_PRESALE",
  REFERRER_CLAIM_PRESALE = "REFERRER_CLAIM_PRESALE",
  PURCHASE_WITH_BUSD = "PURCHASE_WITH_BUSD",
  PURCHASE_WITH_NTR = "PURCHASE_WITH_NTR",
  CLAIM_CENTHER_HISTORY = "CLAIM_CENTHER_HISTORY",
  REGISTRATION_HISTORY = "REGISTRATION_HISTORY",
}